pyproject.toml
[tool.poetry]
name = "fluxcd-teams-bot"
version = "0.1.0"
description = ""
authors = ["Terry Cain <terry@terrys-home.co.uk>"]

[tool.poetry.dependencies]
python = "^3.8"
botbuilder-core = ">=4.7.0"
aiohttp = "^3.6.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
